+++ lib/domain/record/components/supports/components/display_number_setting/display_number_setting_sheet.dart
@@ -8,7 +8,7 @@ import 'package:pilll/components/atoms/buttons.dart';
 import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/components/atoms/font.dart';
 import 'package:pilll/components/atoms/text_color.dart';
-import 'package:pilll/database/batch.dart';
+import 'package:pilll/provider/batch.dart';
 import 'package:pilll/database/pill_sheet_modified_history.dart';
 import 'package:pilll/entity/pill_sheet_group.codegen.dart';
 import 'package:pilll/provider/pill_sheet_group.dart';
