+++ lib/domain/record/components/supports/components/rest_duration/end_manual_rest_duration_button.dart
@@ -5,7 +5,7 @@ import 'package:pilll/components/atoms/buttons.dart';
 import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/components/atoms/font.dart';
 import 'package:pilll/components/atoms/text_color.dart';
-import 'package:pilll/database/batch.dart';
+import 'package:pilll/provider/batch.dart';
 import 'package:pilll/database/pill_sheet_modified_history.dart';
 import 'package:pilll/entity/pill_sheet.codegen.dart';
 import 'package:pilll/entity/pill_sheet_group.codegen.dart';
