+++ test/domain/initial_setting/initial_setting_store_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/services.dart';
-import 'package:pilll/database/batch.dart';
+import 'package:pilll/provider/batch.dart';
 import 'package:pilll/domain/initial_setting/initial_setting_state_notifier.dart';
 import 'package:pilll/entity/pill_sheet.codegen.dart';
 import 'package:pilll/entity/pill_sheet_group.codegen.dart';
